% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_ratios.R
\name{abundance_ratios}
\alias{abundance_ratios}
\title{Calculate abundance ratio based on input intensities}
\usage{
abundance_ratios(
  .data,
  ...,
  i44 = s44,
  i45 = s45,
  i46 = s46,
  i47 = s47,
  i48 = s48,
  i49 = s49,
  R45 = R45,
  R46 = R46,
  R47 = R47,
  R48 = R48,
  R49 = R49,
  quiet = NULL
)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{i44, i45, i46, i47, i48, i49}{Name of mass x column. Defaults to \code{sx}.}

\item{R45, R46, R47, R48, R49}{Desired new name of the calculated ratio for mass x. Defaults to \code{Rx}.}

\item{quiet}{If \code{TRUE}, print info messages. If \code{NULL} (default) listens to \code{options(clumpedr.quiet)}}
}
\value{
Same as \code{.data} but with new columns Rx.
}
\description{
Calculate abundance ratios by dividing input mass column \eqn{i} over mass
44. \deqn{R_i = mass_i / mass_{44}}
}
\examples{
# easily generate some artificial data
x <- function() rnorm(n = 10)
dat <- tibble(s44 = x(), s45 = x(), s46 = x(), s47 = x(), s48 = x(), s49 = x())
dat <- dat |>
  abundance_ratios()
# and in case you wish to specify the column names (e.g. when we calculate the bg ratios)
dat <- dat |>
  abundance_ratios(i44 = r44, i45 = r45, i46 = r46, i47 = r47, i48 = r48, i49 = r49,
                  R45 = R45_wg, R46 = R46_wg, R47 = R47_wg, R48 = R48_wg, R49 = R49_wg)
}
