% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info.R
\name{add_info}
\alias{add_info}
\title{Add the information from the did files}
\usage{
add_info(.data, .info, ..., cols = NULL, quiet = NULL)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble}, resulting from
\code{\link[=collapse_cycles]{collapse_cycles()}}.}

\item{.info}{A \link[tibble:tibble-package]{tibble}, resulting from
\code{\link[=clean_did_info]{clean_did_info()}}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{cols}{A character vector with column names in info to add to the data.}

\item{quiet}{If \code{TRUE}, print info messages. If \code{NULL} (default) listens to \code{options(clumpedr.quiet)}}
}
\value{
Same as \code{.data} but with new columns in \code{cols}.
}
\description{
Append mass spectrometer metadata to the raw data.
}
\examples{
# generate some artificial data
dat <- tibble(file_id = c("hi", "bye"), Analysis = c(1, 2))
# and artificial metadata
inf <- tibble(file_id = c("hi", "bye"), Analysis = c(1, 2),
              file_path = c("/path/to/file", "/path/to/file"),
              weight = c(69, 87), )
dat <- dat |>
  add_info(inf, c("weight", "file_path"))
}
\seealso{
Other metadata cleaning functions: 
\code{\link{clean_did_info}()},
\code{\link{get_inits}()},
\code{\link{parse_info}()}
}
\concept{metadata cleaning functions}
