% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_expected_values.R
\name{append_expected_values}
\alias{append_expected_values}
\title{Append expected values}
\usage{
append_expected_values(
  .data,
  ...,
  std_names = paste0("ETH-", 1:3),
  std_values = c(0.2052, 0.2085, 0.6132),
  exp = expected_D47,
  by = `Identifier 1`,
  quiet = NULL
)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{std_names}{Names of the standards.}

\item{std_values}{Expected values of the standards. Defaults to Bernasconi et al., 2021.}

\item{exp}{Name of the new column that will hold expected values.}

\item{by}{Name of the standard/sample identifier column.}

\item{quiet}{If \code{TRUE}, print info messages. If \code{NULL} (default) listens to \code{options(clumpedr.quiet)}}
}
\value{
Same as \code{.data} but with new column in \code{exp}.
}
\description{
Append the expected values for the standards. Defaults to using
ETH-1--ETH-3.
}
\examples{
standards |>
  isoreader::iso_get_raw_data(include_file_info = "Identifier 1") |>
  append_expected_values(std_names = c("ETH-1", "ETH-2", "ETH-3"),
                         std_values = c(0.2052, 0.2085, 0.6132),
                         exp = expected_D47,
                         by = `Identifier 1`)
}
\references{
Bernasconi, S. M., Daëron, M., Bergmann, K. D., Bonifacie, M., Meckler, A.
N., Affek, H. P., Anderson, N., Bajnai, D., Barkan, E., Beverly, E.,
Blamart, D., Burgener, L., Calmels, D., Chaduteau, C., Clog, M.,
Davidheiser-Kroll, B., Davies, A., Dux, F., Eiler, J., … Ziegler, M.
(2021). InterCarb: A community effort to improve inter-laboratory
standardization of the carbonate clumped isotope thermometer using
carbonate standards. Geochemistry, Geophysics, Geosystems, 22.
\url{https://doi.org/10.1029/2020GC009588}
}
\seealso{
Other empirical transfer functions: 
\code{\link{apply_etf}()},
\code{\link{calculate_etf}()},
\code{\link{empirical_transfer_function}()}
}
\concept{empirical transfer functions}
