% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_ms_info.R
\name{append_ms_info}
\alias{append_ms_info}
\title{Append Mass Spectrometer name and some metadata}
\usage{
append_ms_info(dat, masspec, standards = paste0("ETH-", 1:4),
  other_label = "other", ms_name = quo(masspec),
  broadid_name = quo(broadid), group = TRUE, sort = TRUE)
}
\arguments{
\item{dat}{The data in flattened format.}

\item{masspec}{The name of the mass spectrometer to be added.}

\item{standards}{The names of the standards, identified in column 'Identifier 1'.}

\item{other_label}{The label name to use for anything not in the standards.}

\item{ms_name}{The column name to use for the mass spec label.}

\item{broadid_name}{The column name for the broad id (standard name or other).}

\item{group}{Logical. Group by file_path, file_datetime, Identifier 1, type if \code{TRUE} (default).}

\item{sort}{Logical. Arrange by the newly created groups if \code{TRUE} (default).}
}
\description{
Append mass spectrometer name, as well as a new "broadid" for easy plotting.
TODO: Furthermore, add an v44.mV_init column
}
