% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_gas_delta.R
\name{append_ref_deltas}
\alias{append_ref_deltas}
\title{Append reference gas delta values}
\usage{
append_ref_deltas(
  data,
  did = NULL,
  d13C_PDB_wg = NULL,
  d18O_PDBCO2_wg = NULL,
  quiet = default(quiet)
)
}
\arguments{
\item{data}{A \link[tibble:tibble-package]{tibble} to append the delta values to.}

\item{did}{An iso file, resulting from \code{\link[isoreader:iso_read_dual_inlet]{isoreader::iso_read_dual_inlet()}}.}

\item{d13C_PDB_wg}{\eqn{\delta^{13}C} reference gas value to overwrite.}

\item{d18O_PDBCO2_wg}{\eqn{\delta^{18}O} reference gas value to overwrite.}
}
\description{
Append reference gas delta values
}
