% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_and_clumping_deltas.R
\name{bulk_and_clumping_deltas}
\alias{bulk_and_clumping_deltas}
\title{Compute δ^13^C (‰ PDB), δ^18^O (‰ PDB-CO2), Δ~47~, Δ~48~ and Δ~49~ values
(‰)}
\usage{
bulk_and_clumping_deltas(dat, d13C_PDB_wg, d18O_PDBCO2_wg,
  d45 = quo(d45), d46 = quo(d46), d47 = quo(d47), d48 = quo(d48),
  d49 = quo(d49), R13_PDB = default(R13_PDB),
  R18_PDB = default(R18_PDB), R17_PDBCO2 = default(R17_PDBCO2),
  R18_PDBCO2 = default(R18_PDBCO2), lambda = default(lambda),
  D17O = default(D17O), quiet = default(quiet))
}
\arguments{
\item{dat}{A tibble containing delta values 45 through 49.}

\item{d13C_PDB_wg}{Reference gas δ^13^C value.}

\item{d18O_PDBCO2_wg}{Reference gas δ^13^C gas value.}

\item{R17_PDBCO2}{The R17 value of PDB (g). Defaults to 0.0003931, from
Assonov and Brenninkmeijer 2003, re-scaled by Brand, 2010.}

\item{R18_PDBCO2}{The R18 value of PDB (g). Defaults to 0.00208839, derrived
from Baertschi, 1976, calculated by Brand 2010 eq. 2.}

\item{lambda}{Isotopic fractionation of ^17^O. Defaults to 0.528, from
Barkan and Luz, 2005.}

\item{D17O}{TODO: look up in Daëron 2016 what this does again.}
}
\description{
Requires a dataframe with little δ values, and working gas δ^13^C
and δ^18^O values.
}
