% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpedr-package.R
\docType{package}
\name{clumpedr-package}
\alias{clumpedr}
\alias{clumpedr-package}
\title{A package for clumped isotope analysis in R}
\description{
This package makes analyzing clumped isotope data as easy as possible.
}
\details{
This package incorporates functions to optimize the distribution of
(carbonate) standards and the optimal proportion of standards versus the
number of sample replicates. In the future, the package will include further
scripts for clumped isotope analysis.

TODO: include vignettes
}
\section{Package options}{

The functions in this package listen to the \code{genplot} and
\code{quiet} options, which can be set using
}

\author{
Ilja J. Kocken, \email{i.j.kocken@uu.nl}
}
\keyword{internal}
