% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{clumpedr_get_default_parameters}
\alias{clumpedr_get_default_parameters}
\title{Get the current default parameters}
\usage{
clumpedr_get_default_parameters()
}
\description{
Retrieve a table with all default function parameters for this package. To
set read parameters, see \code{\link{clumpedr_set_default_read_parameters}}.
TODO: add clumpedr_set_default_parameters() To set messaging and caching
parameters see \code{\link{clumpedr_info_messages}} and see. For a piping
compatible version of this function, see
\link{clumpedr_show_default_reader_parameters}.
TODO: create show_default_reader_parameters function
}
\seealso{
Other settings functions: \code{\link{clumpedr_info_messages}}
}
\concept{settings functions}
