% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{clumpedr_info_messages}
\alias{clumpedr_info_messages}
\alias{clumpedr_turn_info_messages_on}
\alias{clumpedr_turn_info_messages_off}
\title{Turn information messages on/off}
\usage{
clumpedr_turn_info_messages_on(data = NULL)

clumpedr_turn_info_messages_off(data = NULL)
}
\description{
These functions turn information messages on/off in all subsequent function
calls by changing the global settings for the \code{quiet} parameter of most
clumpedr functions. These functions can be called stand alone or within a
pipeline to turn messages on/off at a certain point during the pipeline.
}
\seealso{
Other settings functions: 
\code{\link{clumpedr_get_default_parameters}()},
\code{\link{clumpedr_show_default_parameters}()}
}
\concept{settings functions}
