% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_backgrounds.R
\name{correct_backgrounds}
\alias{correct_backgrounds}
\title{Background corrections}
\usage{
correct_backgrounds(.data, factor, i47 = v47.mV, i54 = v54.mV,
  quiet = default(quiet))
}
\arguments{
\item{.data}{The dataframe with raw iso files.}

\item{factor}{Factor by which to multiply the half-cup before subtraction.}

\item{i47}{Column with mass 47 intensities to correct.}

\item{i54}{Column with mass 47.5 intensities to use for correction.}
}
\description{
This function applies a very simple background correction based on the cup
54 intensity and an input scaling factor.
}
