% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{default}
\alias{default}
\title{retrieve package settings}
\usage{
default(name)
}
\description{
first checks if the setting exists in the isoreader space, then in the isoprocessor
}
\details{
Consider providing an option that indicates whether to return the expression or a default value
e.g. default(x) that returns expr(x) if x is not set vs. default(x, NULL) that returns expr(NULL) if x is not set
alternatively set ALL available parameters in the initialize_options and force parameters to exist
}
