% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disable_cycles.R
\name{disable_cycles}
\alias{disable_cycles}
\title{Disable cycles that failed}
\usage{
disable_cycles(dat, min = 1500, max = 50000, fac = 1.5,
  m44.mV = quo(v44.mV), cycle = quo(cycle), relative_to = "init",
  genplot = default(genplot), quiet = default(quiet))
}
\arguments{
\item{dat}{A \link[tibble:tibble-package]{tibble}, resulting from
\code{isoreader::iso_get_raw_data()}}

\item{min}{Minimum intensity level for good cycles. Defaults to 1,500 mV.}

\item{max}{Maximum intensity level for good cycles. Defaults to 50,000 mV.}

\item{fac}{Factor for how much larger the current drop should be than the
one specified in \code{relative_to}.}

\item{cycle}{quoted column name of the column with the measurement cycle
number.}

\item{relative_to}{cycle Drop detection occurs relative to either the first
cycle ("init", default) or to the previous cycle ("prev").}
}
\description{
Based on several criteria, filter out cycles that show a sudden pressure drop.
}
\details{
This function tries to identify cycles in which the pressure has
dropped fast. It first disables all the cycles above or below certain
thresholds (min, max), and then TODO
}
\seealso{
Other cycle functions: \code{\link{find_bad_cycles}},
  \code{\link{plot_disabled_cycles}}
}
\concept{cycle functions}
