% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eth3}
\alias{eth3}
\title{Single ETH-3 standard did file}
\format{
An object of class \code{dual_inlet} (inherits from \code{iso_file}) of length 7.
}
\usage{
eth3
}
\description{
Single ETH-3 standard did file
}
\keyword{datasets}
