% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_clumped.R
\name{filter_clumped}
\alias{filter_clumped}
\title{Filter out all the clumped files from regular stable isotope geochemistry}
\usage{
filter_clumped(.did, regex = NULL, quiet = default(quiet))
}
\arguments{
\item{.did}{A \link[tibble:tibble-package]{tibble}, resulting from
\link[isoreader:iso_read_dual_inlet]{iso_read_dual_inlet()}}

\item{regex}{The regular expression that captures the desired methods. The
default is a very explicit filtering of our presently used method names,
be sure to change this!}
}
\description{
This function applies the \link[isoreader:iso_filter_files]{iso_filter_files()}
function to filter by the Method using a regular expression.
}
\references{
\url{https://en.wikipedia.org/wiki/Regular_expression}
}
