% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_init_outliers}
\alias{find_init_outliers}
\title{Find measurements that have wrong initial intensities.}
\usage{
find_init_outliers(.data, init_low = 8000, init_high = 40000,
  init_diff = 1200, quiet = default(quiet))
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble} with raw Delta values and file information.}

\item{init_low}{Minimum initial intensity threshold for mass 44.}

\item{init_high}{Maximum initial intensity threshold for mass 44.}

\item{init_diff}{Maximum initial difference in mass 44 threshold between standard and sample gas.}
}
\description{
Finds values of \code{s44_init} and \code{r44_init} that fall outside of
\code{init_low} and \code{init_high}, as well as those where the difference
between the reference gas intensity and the sample gass intensity is larger
than \code{init_diff}.
}
