% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_internal_sd_outlier}
\alias{find_internal_sd_outlier}
\title{Find internal standard deviation outliers.}
\usage{
find_internal_sd_outlier(
  .data,
  internal_sd = 0.15,
  ...,
  D47 = D47_raw,
  quiet = default(quiet)
)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble} with raw Delta values and file information.}

\item{internal_sd}{The internal standard deviation cutoff value.}

\item{D47}{The column to calculate the internal sd value for.}
}
\description{
Outliers are flagged based on an internal standard deviation cutoff value
\code{internal_sd}.
}
