% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_outliers}
\alias{find_outliers}
\title{Find outliers}
\usage{
find_outliers(dat, init = 8000, diff = 1200, nsd_off = 4,
  std_names = paste0("ETH-", 1:3))
}
\arguments{
\item{dat}{A \link[tibble:tibble-package]{tibble}.}

\item{parameter}{parameter description}
}
\description{
This function finds outliers based on several criteria.
}
\details{
Here, we define an outlier as a measurement that has:
\itemize{
\item an initial mass 44 intensity below \code{init}.
\item an imbalance between sample and reference gas of more than \code{diff}.
\item a clumped value that is more than \code{nsd_off} standard deviations away from the mean.
}
}
