% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_param49_outliers}
\alias{find_param49_outliers}
\title{Find param 49 outliers.}
\usage{
find_param49_outliers(.data, param49_off, ..., quiet = NULL)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble} with raw Delta values and file information.}

\item{param49_off}{The absolute cutoff value for the parameter 49 value.}

\item{...}{These dots are for future extensions and must be empty.}

\item{quiet}{If \code{TRUE}, print info messages. If \code{NULL} (default) listens to \code{options(clumpedr.quiet)}}
}
\description{
Find measurements with a param49 value that is greater than \code{param49_off}.
}
