% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_R_flags}
\alias{find_R_flags}
\title{Find R45 and R46 outliers.}
\usage{
find_R_flags(.data, quiet = default(quiet))
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble} with raw Delta values and file information.}
}
\description{
If (R45 / R45_stoch - 1) > 2e-8, we flag it as an outlier.
}
