% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_session_id1_outlier}
\alias{find_session_id1_outlier}
\title{Find session outliers by sample/standard type.}
\usage{
find_session_id1_outlier(
  .data,
  n_id1 = 5,
  nsd_off = 4,
  D47 = D47_raw,
  session = Preparation,
  id1 = `Identifier 1`,
  quiet = default(quiet)
)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble} with raw Delta values and file information.}

\item{n_id1}{The minimum number of measurements of the sample/standard in the session needed to calculate an offset from the median.}

\item{nsd_off}{The number of standard deviations away from the median.}

\item{D47}{The column to calculate the internal sd value for.}

\item{session}{The session for which to calculate the standard deviation and median values.}

\item{id1}{The column that defines the sample/standard name.}
}
\description{
Find outliers that are more than 4 standard deviations away from the session and sample/standard median.
}
