% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_did_info.R
\name{get_inits}
\alias{get_inits}
\title{Add initial intensities}
\usage{
get_inits(did)
}
\description{
Add initial intensities
}
\seealso{
Other metadata cleaning functions: \code{\link{add_info}},
  \code{\link{clean_did_info}}, \code{\link{parse_info}}
}
\concept{metadata cleaning functions}
