% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref_delta.R
\name{get_ref_delta}
\alias{get_ref_delta}
\title{Get reference gas delta values}
\usage{
get_ref_delta(.did)
}
\arguments{
\item{.did}{An iso file, resulting from \code{\link[isoreader:iso_read_dual_inlet]{isoreader::iso_read_dual_inlet()}}.}
}
\description{
Get reference gas delta values
}
