% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{kaff}
\alias{kaff}
\title{kaff}
\format{An object of class \code{numeric} of length 1.}
\usage{
kaff
}
\description{
Acid fractionation factor
}
\details{
0.064 from Meckler et al. 2014, 0.062 from Müller et al. 2017
}
\keyword{datasets}
