% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{kintercept}
\alias{kintercept}
\title{kintercept}
\format{An object of class \code{numeric} of length 1.}
\usage{
kintercept
}
\description{
Intercept of our long-term empirical transfer function
raw values ~ expected values
}
\seealso{
kslope
}
\keyword{datasets}
