% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{klab_smpid}
\alias{klab_smpid}
\title{klab_smpid}
\format{An object of class \code{character} of length 2.}
\usage{
klab_smpid
}
\description{
Values of sample id legend, c("Cold (0°C)", "Warm (40°C)")
}
\keyword{datasets}
