% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{krev.slope}
\alias{krev.slope}
\title{krev.slope}
\format{An object of class \code{numeric} of length 1.}
\usage{
krev.slope
}
\description{
Inverse slope of our long-term empirical transfer function
expected values ~ raw values
}
\seealso{
krev.intercept
}
\keyword{datasets}
