% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{kslope}
\alias{kslope}
\title{kslope}
\format{An object of class \code{numeric} of length 1.}
\usage{
kslope
}
\description{
Slope of our long-term emperical transfer function
raw values ~ expected values
}
\seealso{
kintercept
}
\keyword{datasets}
