% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ktit_smpid}
\alias{ktit_smpid}
\title{ktit_smpid}
\format{An object of class \code{character} of length 1.}
\usage{
ktit_smpid
}
\description{
Title of sample id legend, "Sample at temperature"
}
\keyword{datasets}
