% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_std_table.R
\name{make_std_table}
\alias{make_std_table}
\title{Create reference table of standards}
\usage{
make_std_table(id = c(paste0("ETH-", 1:4), "UU"), col = c("orange",
  "purple", "#00B600", "blue", "#FFCD00"), D47_std = c(0.258, 0.256,
  0.691, 0.507, tempcal(4, ignorecnf = TRUE)), D47_95 = c(0.003, 0.004,
  0.004, 0.004, 0.004), slope = kslope, intercept = kintercept,
  aff = kaff)
}
\arguments{
\item{id}{The names of the standards that are used in the simulations.}

\item{col}{The colours that these standards will have in plots.}

\item{D47_std}{The accepted Δ47 values of the standards.}
}
\value{
A tibble of standard names (id), colours (col), Δ47 accepted value
(D47), Δ47 with acid fractionation subtracted, i.e. projected to 70°C
(D47.noacid), the expected raw value based on the linear regression
(rawcat) and the temperature in °C (temp).
}
\description{
Create a reference table with expected values for the standards
}
\seealso{
make_smp_info kslope kintercept kaff
}
