% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_intensities.R
\name{match_intensities}
\alias{match_intensities}
\title{Match the reference gas intensity to that of the sample gas}
\usage{
match_intensities(dat, method = "normal", quiet = default(quiet))
}
\arguments{
\item{dat}{A dataframe with s44--s49 and r44--r49, output of
\code{spread_intensities()}.}

\item{method}{"linterp" for linear interpolation, or "normal" for
conventional bracketing of sample gas.}
}
\description{
This matches the reference gas intensity of mass 44 to the sample gas
intensity of mass 44 through linear interpolation (option \code{method = "linterp"}), and then applies this same offset to the other masses.
}
