% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_cycles.R
\name{nest_cycle_data}
\alias{nest_cycle_data}
\title{nest cycle data}
\usage{
nest_cycle_data(
  .data,
  ...,
  masses = c(44:49, 54),
  ratios = c(paste0("s", masses), paste0("r", masses)),
  outlier_cycles = c(paste0("outlier_cycle_", c("low_s44", "low_r44", "high_s44",
    "high_r44", "drop_s44", "drop_r44", "s44", "r44")), "outlier_cycle"),
  outliers = "outlier",
  cycle_drop = paste0("cycle_", c("diff_s44", "diff_r44", "drop_s44", "drop_r44",
    "drop_num_s44", "drop_num_r44", "has_drop_s44", "has_drop_r44")),
  bgs = c(paste0("s44_bg", masses[!masses \%in\% c(44, 54)]), paste0("r44_bg",
    masses[!masses \%in\% c(44, 54)])),
  bg_corrected = paste0(ratios[!ratios \%in\% c("s44", "r44", "r54", "s54")], "_bg"),
  Rs = c("R45", "R46", "R47", "R48", "R49", "R45_wg", "R46_wg", "R47_wg", "R48_wg",
    "R49_wg", "R13_wg", "R18_wg", "R17", "R18", "R13"),
  deltas = c("d13C_PDB", "d18O_PDBCO2", "d18O_PDB", "d45", "d46", "d47", "d48", "d49"),
  isotopes = c("C12", "C13", "C16", "C17", "C18", "C626", "C627", "C628", "C636", "C637",
    "C638", "C727", "C728", "C737", "C738", "C828", "C838"),
  params = c("K", "A", "B", "C", "D", "aa", "bb", "cc"),
  stochastic = c("R45_stoch", "R46_stoch", "R47_stoch", "R48_stoch", "R49_stoch"),
  flags = c("R45_flag", "R46_flag"),
  Deltas = c("D45_raw", "D46_raw", "D47_raw", "D48_raw", "D49_raw"),
  p49 = "param_49",
  quiet = NULL
)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble} resulting from \code{\link[=bulk_and_clumping_deltas]{bulk_and_clumping_deltas()}}.}

\item{...}{Additional chacter vectors to nest by.}

\item{masses}{The masses that are present in the mass spectrometer. Defaults to 44:49 and 54.}

\item{ratios}{Ratio columns, based on masses.}

\item{outlier_cycles}{Columns with outlier_cycle information.}

\item{outliers}{Other outlier-related columns.}

\item{cycle_drop}{Columns related to detecing cycle drops.}

\item{bgs}{Backgrounds.}

\item{bg_corrected}{Background-corrected raw intensities per cup.}

\item{Rs}{R values.}

\item{deltas}{delta values.}

\item{isotopes}{C values.}

\item{params}{Taylor polynomial parameters.}

\item{stochastic}{Stochastic R values.}

\item{flags}{Flag columns.}

\item{Deltas}{Big delta values.}

\item{p49}{Param 49.}

\item{quiet}{If \code{TRUE}, print info messages. If \code{NULL} (default) listens to \code{options(clumpedr.quiet)}}
}
\description{
Nest the cycle data resulting from previous computations steps.
}
\details{
The parameters are all strings with column names. They default to all the
columns that have cycle information based on previous computation steps.
}
