% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_base.R
\name{plot_base}
\alias{plot_base}
\title{plot_base}
\usage{
plot_base(dat, ...)
}
\arguments{
\item{dat}{A \link[tibble:tibble-package]{tibble}, resulting from any of the
processing functions that returns the metadata.}

\item{...}{Additional aesthetics to pass to ggplot.}
}
\description{
This function creates a ggplot with all of the extra default aesthetics set up.
}
