% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_delta.R
\name{plot_delta}
\alias{plot_delta}
\title{plot_delta}
\usage{
plot_delta(dat, col = quo(D47_final))
}
\arguments{
\item{dat}{A \link[tibble:tibble-package]{tibble}, resulting from \code{temperature_calculation}}

\item{col}{Target column.}
}
\description{
plot_delta
}
