% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_disabled_cycles.R
\name{plot_disabled_cycles}
\alias{plot_disabled_cycles}
\title{Plot disabled cycles}
\usage{
plot_disabled_cycles(dat, min = 1500, max = 50000,
  quiet = default(quiet))
}
\arguments{
\item{dat}{Dataframe with output from \code{disable_cycles()}.}
}
\description{
Plot disabled cycles
}
