% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_transfer_function.R
\name{plot_etf}
\alias{plot_etf}
\title{Plot the Empirical Transfer Function}
\usage{
plot_etf(dat, std_names = paste0("ETH-", 1:3),
  session = quo(Preparation))
}
\arguments{
\item{session}{The quoted column name to group analyses by. Defaults to \code{quo(Preparation)}.}
}
\description{
Create a plot with the expected \eqn{\Delta_{47}}{Δ47} value on the x-axis
and the \eqn{\Delta_{47}}{Δ47} derrived from the ETF on the y-axis.
}
\seealso{
Other empirical transfer functions: \code{\link{append_expected_values}},
  \code{\link{apply_etf}},
  \code{\link{empirical_transfer_function}}
}
\concept{empirical transfer functions}
