% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outliers.R
\name{plot_outliers}
\alias{plot_outliers}
\title{Plot outliers}
\usage{
plot_outliers(dat, col = quo(D47_raw_mean))
}
\arguments{
\item{dat}{A \link[tibble:tibble-package]{tibble}, resulting from \code{\link[=add_info]{add_info()}}}

\item{col}{The quoted column that should be plotted.}
}
\description{
This function plots raw delta values that are outliers, based on several criteria.
}
