% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_delta.R
\name{plot_raw_delta}
\alias{plot_raw_delta}
\title{Plot raw delta values as a function of measurement time.}
\usage{
plot_raw_delta(
  .data,
  .info,
  x = file_datetime,
  y = D47_raw,
  shape = outlier_cycle,
  raw_points = FALSE,
  point_alpha = 0.5,
  quiet = default(quiet)
)
}
\arguments{
\item{.data}{The input tibble, result of \code{\link[=bulk_and_clumping_deltas]{bulk_and_clumping_deltas()}}.}

\item{.info}{The metadata of the raw cycles, used to get the \code{file_datetime}.}

\item{x}{X aesthetic. Defaults to file_datetime.}

\item{y}{Y aesthetic. Defaults to D47_raw.}

\item{shape}{Shape aesthetic. Defaults to \code{outlier_cycle}.}

\item{raw_points}{Logical, whether to include individual cycle estimates (defaults to \code{FALSE}).}

\item{point_alpha}{The alpha value of the raw points if they are plotted.}
}
\description{
Cycles are averaged and pointranges indicate one standard deviation offset
based on the number of cycles taken into account.
}
\details{
Note that with \eqn{\delta^{13}}C and \eqn{\delta^{18}}O, the
pointrange is invisible because the cycles have so little variance that
the line segment usually falls within the dots.
}
