% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim.R
\name{plot_sim}
\alias{plot_sim}
\title{Plot output of single simulation}
\usage{
plot_sim(std, smp, cis, stdtable, smpinfo, rawstdev)
}
\arguments{
\item{std}{TODO}

\item{smp}{TODO}

\item{cis}{TODO}

\item{stdtable}{TODO}

\item{smpinfo}{TODO}

\item{stdev}{TODO}
}
\value{
A ggplot object.
}
\description{
This function TODO
}
\references{
Brandon M. Greenwell and Christine M. Schubert Kabban (2014).
investr: An R Package for Inverse Estimation. The R Journal, 6(1),
90-100.
\url{https://journal.r-project.org/archive/2014-1/greenwell-kabban.pdf}
}
