% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stds.R
\name{plot_stds}
\alias{plot_stds}
\title{Plot standards}
\usage{
plot_stds(dat, x = d47_mean, y = D47_final, ...)
}
\arguments{
\item{dat}{A \link[tibble:tibble-package]{tibble}, resulting from \code{\link[=acid_fractionation]{acid_fractionation()}}}
}
\description{
Plot the standards in the familiar \eqn{\delta}{δ} vs \eqn{\Delta}{Δ} plot.
}
