% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{set_default}
\alias{set_default}
\title{set package setting}
\usage{
set_default(name, value, overwrite = TRUE)
}
\description{
Essentially the same function as in isoreader except with the isoprocessor. prefix
}
