% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_intensities.R
\name{spread_intensities}
\alias{spread_intensities}
\title{Spread intensities from iso_read format into wide format.}
\usage{
spread_intensities(dat, quiet = default(quiet))
}
\arguments{
\item{dat}{a data frame or tibble containing mass intensities per cycle}
}
\value{
a \link[=tibble:tibble-package]{tibble}  with the sample  and reference
gasses side-by-side.
}
\description{
To be able to calculate ratios per cycle, it is necessary to put the sample
and reference as columns next to each other. This function transforms the
data into the desired format.
}
