% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standards}
\alias{standards}
\title{27 Dual Inlet did files}
\format{
A "dual_inlet_list" "iso_file_list", resulting from isoreader::iso_read_dual_inlet()
}
\usage{
standards
}
\description{
The standards of 27 dual inlet files measured on 2018-08-14.
}
\keyword{datasets}
