% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{tempcal}
\alias{tempcal}
\title{Temperature calibration}
\usage{
tempcal(
  Tc,
  slope = 0.0449,
  intercept = 0.167,
  slpcnf = 0.001,
  intcnf = 0.01,
  ignorecnf = FALSE
)
}
\arguments{
\item{Tc}{The temperature in \eqn{^{\circ}}C.}

\item{slope}{The slope of the regression.}

\item{intercept}{The intercept of the regression.}

\item{slpcnf}{The confidence of the slope of the regression.}

\item{intcnf}{The confidence of the intercept of the regression.}

\item{ignorecnf}{Whether or not to ignore the confidence of the temperature
calibration.}
}
\value{
A \link[tibble:tibble-package]{tibble} with input Tc and estimated
\eqn{\Delta_47} value with lower and upper bounds.
}
\description{
A clumped isotope temperature calibration of the form:
\deqn{y = a \times 10^6 / T^2 + b}{y = a * 10^6 / T^2 + b},
where T is in Kelvin.
}
\details{
Defaults to Bonifacie et al. 2017
\deqn{\Delta_47 = (0.0449 \pm 0.001 \times 10^6) / T^2 + (0.167 \pm 0.01)}
}
\references{
M. Bonifacie, D. Calmels, J. M. Eiler, J. Horita, C. Chaduteau,
C. Vasconcelos, P. Agrinier, A. Katz, B. H. Passey, J. M. Ferry, J.
Bourrand. Calibration of the dolomite clumped isotope thermometer from 25
to 350 \eqn{^\circ}C, and implications for a universal calibration for all (Ca, Mg,
Fe)CO\eqn{{}_3} carbonates. \strong{2017}, \emph{200}, 255--279.
}
\seealso{
revcal()
}
