% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{temperature_axis}
\alias{temperature_axis}
\title{Generate a formula for a secondary temperature y-axis}
\usage{
temperature_axis(slope = 0.0449, intercept = 0.167)
}
\arguments{
\item{slope}{The slope of the calibration.}

\item{intercept}{The intercept of the calibration.}
}
\description{
Uses the clumped isotope temperature calibration of the form:
\deqn{y = a \times 10^6 / T^2 + b}{y = a * 10^6 / T^2 + b}
}
\details{
Defaults to Bonifacie et al. 2017
\deqn{\Delta_47 = (0.0449 \pm 0.001 \times 10^6) / T^2 + (0.167 \pm 0.01)}
}
\examples{
dat <- tibble::tibble(age = 1:10, D47 = rnorm(10, 0.7, .1))
# create an empty plot
ggplot2::ggplot(dat, ggplot2::aes(x = age, y = D47)) +
  ggplot2::geom_point() +
  # generate the formula for temperature on the fly
  ggplot2::scale_y_continuous(sec.axis = ggplot2::sec_axis(temperature_axis()))
}
\seealso{
temperature_calculation()

revcal()

tempcal()
}
