% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_ratios.R
\name{abundance_ratios}
\alias{abundance_ratios}
\title{Calculate abundance ratio based on input intensities}
\usage{
abundance_ratios(dat, i44 = quo(s44), i45 = quo(s45), i46 = quo(s46),
  i47 = quo(s47), i48 = quo(s48), i49 = quo(s49), R45 = quo(R45),
  R46 = quo(R46), R47 = quo(R47), R48 = quo(R48), R49 = quo(R49))
}
\arguments{
\item{dat}{Input dataframe.}

\item{i44}{Name of mass 44 column. Same for i45--i49.}

\item{R45}{Desired new name of the calculated ratio for mass 45. Same for
R46--R49.}
}
\description{
Calculate abundance ratio based on input intensities
}
