% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acid_fractionation.R
\name{acid_fractionation}
\alias{acid_fractionation}
\title{Add acid fractionation}
\usage{
acid_fractionation(dat, aff = 0.062, quiet = default(quiet))
}
\arguments{
\item{dat}{Data.}

\item{aff}{Temperature-dependent acid fractionation projection from 70 °C to
25 °C. Defaults to 0.062, which is the average of the values obtained by
De Vlieze et al., 2015 and Murray et al., 2016. See also Müller et al.,
2017.}
}
\description{
Add acid fractionation
}
