% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acid_fractionation.R
\name{acid_fractionation}
\alias{acid_fractionation}
\title{Add acid fractionation}
\usage{
acid_fractionation(
  .data,
  aff = 0.062,
  D47 = D47_etf,
  D47_out = D47_final,
  quiet = default(quiet)
)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble}.}

\item{aff}{Temperature-dependent acid fractionation projection from 70 °C to
25 °C. Defaults to 0.062, which is the average of the values obtained by
De Vlieze et al., 2015 and Murray et al., 2016. See Müller et al., 2017.}

\item{D47}{The column name of the \eqn{\Delta_47} values to use for the acid
fractionation calculation.}

\item{D47_out}{The desired new column name.}
}
\description{
The temperature-dependent acid fractionation projection from 70 °C to 25 °C.
It subtracts the \code{aff} from the \code{D47}-value.
}
\references{
W. F. Defliese, M.T. Hren, K. C. Lohmann. Compositional and temperature
effects of phosphoric acid fractionation on \eqn{\Delta_{47}} analysis and
implications for discrepant calibrations. \emph{Chem. Geol.} \strong{2016}, \emph{396}, 51.

S. T. Murray, M. M. Arienzo, P. K. Swart. Determining the \eqn{\Delta_{47}}
acid fractionation in dolomites. \emph{Geochem. Cosmochim. Acta} \strong{2016}, \emph{174},
42.

I. A. Müller, A. Fernandez, J. Radke, J. van Dijk, D. Bowen, J. Schwieters,
S. M. Bernasconi. Carbonate clumped isotope analyses with the
long-integration dual-inlet (LIDI) workflow: scratching at the lower sample
weight boundaries. \emph{Rapid Commun. Mass Spectrom.} \strong{2017}, \emph{31},
1057--1066.
}
