% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info.R
\name{add_info}
\alias{add_info}
\title{add_info}
\usage{
add_info(dat, info, ..., quiet = default(quiet))
}
\arguments{
\item{dat}{A \link[tibble:tibble-package]{tibble}, resulting from \code{collapse_cycles}}

\item{info}{A \link[tibble:tibble-package]{tibble}, resulting from \code{clean_did_info()}}

\item{...}{Optional additional arguments to pass to the \code{left_join} call.}
}
\description{
Add the info from the did files.
}
