% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info.R
\name{add_info}
\alias{add_info}
\title{Add the information from the did files}
\usage{
add_info(.data, .info, ..., cols = NULL, quiet = NULL)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble}, resulting from
\code{\link[=collapse_cycles]{collapse_cycles()}}.}

\item{.info}{A \link[tibble:tibble-package]{tibble}, resulting from
\code{\link[=clean_did_info]{clean_did_info()}}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{cols}{A character vector with column names in info to add to the data.}

\item{quiet}{If \code{TRUE}, print info messages. If \code{NULL} (default) listens to \code{options(clumpedr.quiet)}}
}
\description{
Append mass spectrometer metadata to the raw data.
}
\seealso{
Other metadata cleaning functions: 
\code{\link{clean_did_info}()},
\code{\link{get_inits}()},
\code{\link{parse_info}()}
}
\concept{metadata cleaning functions}
