% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_transfer_function.R
\name{append_expected_values}
\alias{append_expected_values}
\title{Append expected values}
\usage{
append_expected_values(dat, names = paste0("ETH-", 1:3), D47 = c(0.258,
  0.256, 0.691), aff = 0.062)
}
\arguments{
\item{dat}{Data yadayada.}

\item{names}{Names of the standards.}

\item{D47}{Expected values of the standards at 25 °C, from Müller et al., 2017.}

\item{aff}{Temperature-dependent acid fractionation projection from 70 °C to
25 °C. Defaults to 0.062, which is the average of the values obtained by
De Vlieze et al., 2015 and Murray et al., 2016. See also Müller et al.,
2017.}
}
\description{
Append expected values
}
