% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_expected_values.R
\name{append_expected_values}
\alias{append_expected_values}
\title{Append expected values}
\usage{
append_expected_values(
  .data,
  ...,
  std_names = paste0("ETH-", 1:3),
  std_values = c(0.258, 0.256, 0.691) - 0.062,
  exp = expected_D47,
  by = `Identifier 1`,
  quiet = NULL
)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{std_names}{Names of the standards.}

\item{std_values}{Expected values of the standards. Defaults to Muller et
al., 2017 at 70 \eqn{^\circ}C acid digestion.}

\item{exp}{Name of the new column that will hold expected values.}

\item{by}{Name of the standard/sample identifier column.}

\item{quiet}{If \code{TRUE}, print info messages. If \code{NULL} (default) listens to \code{options(clumpedr.quiet)}}
}
\description{
Append the expected values for the standards. Defaults to using
ETH-1--ETH-3, subtracting the acid fractionation factor before-hand.
}
\references{
W. F. Defliese, M.T. Hren, K. C. Lohmann. Compositional and temperature
effects of phosphoric acid fractionation on \eqn{\Delta_{47}} analysis
and implications for discrepant calibrations. \emph{Chem. Geol.} \strong{2016}, \emph{396},
51.

S. T. Murray, M. M. Arienzo, P. K. Swart. Determining the
\eqn{\Delta_{47}} acid fractionation in dolomites. \emph{Geochem. Cosmochim.
Acta} \strong{2016}, \emph{174}, 42.

I. A. Muller, A. Fernandez, J. Radke, J. van Dijk, D. Bowen, J. Schwieters,
S. M. Bernasconi. Carbonate clumped isotope analyses with the
long-integration dual-inlet (LIDI) workflow: scratching at the lower sample
weight boundaries. \emph{Rapid Commun. Mass Spectrom.} \strong{2017}, \emph{31},
1057--1066.
}
\seealso{
Other empirical transfer functions: 
\code{\link{apply_etf}()},
\code{\link{calculate_etf}()},
\code{\link{empirical_transfer_function}()}
}
\concept{empirical transfer functions}
