% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_transfer_function.R
\name{apply_etf}
\alias{apply_etf}
\title{Apply the ETF}
\usage{
apply_etf(dat, D47 = quo(D47raw_mean))
}
\arguments{
\item{dat}{A \link[tibble:tibble-package]{tibble} containing column D47.}

\item{D47}{The column with \eqn{\Delta_47} values to use.}
}
\description{
Apply the ETF
}
