% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_etf.R
\name{apply_etf}
\alias{apply_etf}
\title{Apply the ETF}
\usage{
apply_etf(
  .data,
  intercept = intercept,
  slope = slope,
  raw = D47_raw,
  out = D47_etf,
  quiet = default(quiet)
)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble} containing column D47.}

\item{intercept}{The column name with the ETF intercept.}

\item{slope}{The column name with the ETF slope.}

\item{raw}{The column with raw values to apply the ETF to.}

\item{out}{The new column name.}
}
\description{
Uses the previously calculated intercept and slope, and uses them to
calculate the \eqn{\Delta_{47}}{Δ47} values.
}
\details{
Note that the intercept and slope were calculated with the
dependent and independent variables in the other direction, so we flip
them here. i.e.: \deqn{\Delta_{47etf} = - (\alpha / \beta) + (1 / \beta)
  \times \Delta_{47_raw}}{Δ47[etf] = - (α / β) + (1 / β) * Δ[47_raw]}
}
\seealso{
Other empirical transfer functions: 
\code{\link{append_expected_values}()},
\code{\link{empirical_transfer_function}()}
}
\concept{empirical transfer functions}
