% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_transfer_function.R
\name{apply_etf}
\alias{apply_etf}
\title{Apply the ETF}
\usage{
apply_etf(dat, D47 = quo(D47_raw_mean))
}
\arguments{
\item{dat}{A \link[tibble:tibble-package]{tibble} containing column D47.}

\item{D47}{The column with \eqn{\Delta_{47}}{Δ47} values to use.}
}
\description{
Uses the previously calculated intercept and slope, and uses them to
calculate the \eqn{\Delta_{47}}{Δ47} values.
}
\details{
Note that the intercept and slope were calculated with the
dependent and independent variables in the other direction, so we flip
them here. i.e.: \deqn{\Delta_{47etf} = - (\alpha / \beta) + (1 / \beta)
  \times \Delta_{47_raw}}{Δ47_etf = - (α / β) + (1 / β) * Δ47_raw}
}
\seealso{
Other empirical transfer functions: \code{\link{append_expected_values}},
  \code{\link{empirical_transfer_function}},
  \code{\link{plot_etf}}
}
\concept{empirical transfer functions}
