% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_and_clumping_deltas.R
\name{bulk_and_clumping_deltas}
\alias{bulk_and_clumping_deltas}
\title{Compute \eqn{\delta^{13}}C, \eqn{\delta^{18}}O \eqn{\Delta_{47}},
\eqn{\Delta_{48}}, and \eqn{\Delta_{49}} values}
\usage{
bulk_and_clumping_deltas(
  .data,
  ...,
  d45 = d45,
  d46 = d46,
  d47 = d47,
  d48 = d48,
  d49 = d49,
  R13_PDB = default(R13_PDB),
  R18_PDB = default(R18_PDB),
  R17_PDBCO2 = default(R17_PDBCO2),
  R18_PDBCO2 = default(R18_PDBCO2),
  lambda = default(lambda),
  D17O = default(D17O),
  quiet = NULL
)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble} containing s44-s49, r44-r49,
R45-R49, R45_wg-R49_wg, d45-d49, and working gas compositions d13C_PDB_wg and
d18OPDBCO2_wg.}

\item{d45}{Column name of d45.}

\item{d46}{Column name of d46.}

\item{d47}{Column name of d47.}

\item{d48}{Column name of d48.}

\item{d49}{Column name of d49.}

\item{R13_PDB}{The R13 value of the PDB reference. Defaults to 0.01118, from
Chang and Li, 1990.}

\item{R18_PDB}{The R18 value of the PDB reference. Defaults to 1.008751,
from Coplen et al., 1983.}

\item{R17_PDBCO2}{The R17 value of PDB (g). Defaults to 0.0003931, from
Assonov and Brenninkmeijer 2003, re-scaled by Brand, 2010.}

\item{R18_PDBCO2}{The R18 value of PDB (g). Defaults to 0.00208839, derrived
from Baertschi, 1976, calculated by Brand 2010 eq. 2.}

\item{lambda}{Isotopic fractionation of \eqn{^{17}}{17}O. Defaults to 0.528,
from Barkan and Luz, 2005.}

\item{D17O}{The difference between the expected \eqn{\delta^{17}}{\u03b417}O
(\eqn{\delta^{18}}{\u03b418}O\eqn{-\lambda*\delta^{18}}{- lambda *\u03b418}O)
and the actual \eqn{\delta^{17}}{\u03b417}O value.}
}
\description{
Requires a dataframe with little \eqn{\delta} values, and working gas
\eqn{\delta^{13}}C and \eqn{\delta^{18}}O values.
}
\references{
Daeron, M., Blamart, D., Peral, M., & Affek, H. P., Absolute
isotopic abundance ratios and the accuracy of \eqn{\Delta_{47}}
measurements, \emph{Chemical Geology} \strong{2016}, \emph{442}, 83--96.
\doi{10.1016/j.chemgeo.2016.08.014}
}
