% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_etf.R
\name{calculate_etf}
\alias{calculate_etf}
\title{Calculate the Empirical Transfer Function}
\usage{
calculate_etf(.data, cycle_dis = cycle_dis, outlier = outlier,
  raw = D47_raw_mean, exp = expected_D47, session = Preparation,
  quiet = default(quiet))
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble}.}

\item{cycle_dis}{Column with disabled cycle information. Looks up string \code{"no_drop"}.}

\item{outlier}{Column with sample outlier information. Looks up string \code{"no_outlier"}.}

\item{raw}{Column name of raw \eqn{\Delta_{47}}{Δ47} values.}

\item{exp}{Column name of expected \eqn{\Delta_{47}}{Δ47} values.}

\item{session}{The column name to group analyses by. Defaults to
\code{Preparation}.}
}
\description{
Calculate the Empirical Transfer Function
}
