% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_transfer_function.R
\name{calculate_etf}
\alias{calculate_etf}
\title{Calculate the Empirical Transfer Function}
\usage{
calculate_etf(.data, session = Preparation, quiet = default(quiet))
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble}.}

\item{session}{The column name to group analyses by. Defaults to
\code{Preparation}.}
}
\description{
Calculate the Empirical Transfer Function
}
