% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_did_info.R
\name{clean_did_info}
\alias{clean_did_info}
\title{Clean up the did file info}
\usage{
clean_did_info(
  .did,
  masspec = NULL,
  ...,
  std_names = c(paste0("ETH-", 1:4), "IAEA-C2", "Merck"),
  oth_name = "other",
  quiet = NULL
)
}
\arguments{
\item{.did}{An iso file, resulting from \code{\link[isoreader:iso_read_dual_inlet]{isoreader::iso_read_dual_inlet()}}.}

\item{masspec}{The name of the mass spectrometer to append.}

\item{...}{These dots are for future extensions and must be empty.}

\item{std_names}{Character vector of the standard names to find in \verb{Identifier 1}.}

\item{oth_name}{Single general name to assign to non-standard "other" measurements.}

\item{quiet}{If \code{TRUE}, print info messages. If \code{NULL} (default) listens to \code{options(clumpedr.quiet)}}
}
\description{
Converts mass spectrometer metadata into the correct types and appends the
initial intensity of mass 44.
}
\seealso{
\code{\link[isoreader:iso_read_dual_inlet]{isoreader::iso_read_dual_inlet()}}

\code{\link[isoreader:iso_get_file_info]{isoreader::iso_get_file_info()}}

Other metadata cleaning functions: 
\code{\link{add_info}()},
\code{\link{get_inits}()},
\code{\link{parse_info}()}
}
\concept{metadata cleaning functions}
