% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_did_info.R
\name{clean_did_info}
\alias{clean_did_info}
\title{Clean up the did file info}
\usage{
clean_did_info(.did, masspec = NULL, std_names = paste0("ETH-", 1:4),
  oth_name = "other", quiet = default(quiet))
}
\arguments{
\item{.did}{An iso file, resulting from \code{\link[isoreader:iso_read_dual_inlet]{isoreader::iso_read_dual_inlet()}}.}

\item{masspec}{The name of the mass spectrometer to append.}

\item{std_names}{Character vector of the standard names to find in \code{Identifier 1}.}

\item{oth_name}{Single general name to assign to non-standard "other" measurements.}
}
\description{
Converts mass spectrometer metadata into the correct types and appends the
initial intensity of mass 44.
}
\seealso{
\code{\link[isoreader:iso_read_dual_inlet]{isoreader::iso_read_dual_inlet()}}

\code{\link[isoreader:iso_get_file_info]{isoreader::iso_get_file_info()}}

Other metadata cleaning functions: \code{\link{add_info}},
  \code{\link{get_inits}}, \code{\link{parse_info}}
}
\concept{metadata cleaning functions}
