% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpedr-package.R
\docType{package}
\name{clumpedr-package}
\alias{clumpedr}
\alias{clumpedr-package}
\title{A package for clumped isotope analysis in R}
\description{
Data processing in clumped isotopes is an involved, evolving field. This package aims facilitate the processing of raw data to final \eqn{\Delta_{47}}{Δ47} values and temperatures, step-by-step, with good documentation, so that the full process is easier to understand and tweak, more reproducible, and fully transparent.
}
\details{
This package incorporates functions to optimize the distribution of
(carbonate) standards and the optimal proportion of standards versus the
number of sample replicates. In the future, the package will include further
scripts for clumped isotope analysis.

TODO: include vignettes
}
\section{Package options}{

The functions in this package listen to the \code{genplot} and
\code{quiet} options, which can be set using
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/isoverse/clumpedr}
  \item Report bugs at \url{https://github.com/isoverse/clumpedr/issues}
}

}
\author{
Ilja J. Kocken, \email{i.j.kocken@uu.nl}
}
\keyword{internal}
