% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_cycles.R
\name{collapse_cycles}
\alias{collapse_cycles}
\title{Collapse the cycles of the raw measurements}
\usage{
collapse_cycles(dat, na.rm = TRUE, quiet = default(quiet))
}
\arguments{
\item{dat}{A tibble resulting from \code{bulk_and_clumping_deltas()}.}

\item{na.rm}{Logical. Whether or not to remove NA values.}
}
\description{
This collapses the (filtered) cycles of the raw measurements and calculates
averages and standard deviations.
}
