% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_backgrounds.R
\name{correct_backgrounds}
\alias{correct_backgrounds}
\title{Background corrections}
\usage{
correct_backgrounds(dat, factor, quiet = default(quiet))
}
\arguments{
\item{dat}{The dataframe with raw iso files.}

\item{factor}{Factor by which to multiply the half-cup before subtraction.}
}
\description{
This function applies a very simple backgorund correction based on the
half-cup mass 54 intensity.
}
