% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default.params}
\alias{default.params}
\alias{clumpedr.params}
\alias{clumpedr.parameters}
\alias{default.parameters}
\alias{brand.parameters}
\alias{Brand}
\alias{brand_parameters}
\title{Default parameters}
\arguments{
\item{R13_PDB}{The R13 value of the PDB reference. Defaults to 0.01118, from
Chang and Li, 1990.}

\item{R18_PDB}{The R18 value of the PDB reference. Defaults to 1.008751,
from Coplen et al., 1983.}

\item{R17_PDBCO2}{The R17 value of PDB (g). Defaults to 0.0003931, from
Assonov and Brenninkmeijer 2003, re-scaled by Brand, 2010.}

\item{R18_PDBCO2}{The R18 value of PDB (g). Defaults to 0.00208839, derrived
from Baertschi, 1976, calculated by Brand 2010 eq. 2.}

\item{lambda}{Isotopic fractionation of \eqn{^{17}}{17}O. Defaults to 0.528,
from Barkan and Luz, 2005.}

\item{D17O}{TODO: look up in Daëron 2016 what this does again.}

\item{D47}{TODO: look up}

\item{D48}{TODO: look up}

\item{D49}{TODO: look up}
}
\description{
Default parameters
}
\references{
Assonov, S., Brenninkmeijer C. On the \eqn{^{17}}{17}O correction for
CO\eqn{_{2}}{2} mass spectrometric isotopic analysis \emph{Rapid Commun. Mass
Spectrom.}, \strong{2003}, 17, pp. 1007-1016, 10.1002/rcm.1012

Baertschi, P., Absolute \eqn{^{18}}{18}O content of standard mean ocean
water \emph{Earth Planet. Sci. Lett.}, \strong{1976}, 31, pp. 341-344

Barkan, E., Luz B. High-precision measurements of
\eqn{^{17}}{17}O/\eqn{^{16}}{16}O and \eqn{^{18}}{18}O/\eqn{^{16}}{16}O
ratios in H2O \emph{Rapid Commun. Mass Spectrom.}, \strong{2005}, 19, pp. 3737-3742

Chang, T., Li W. A calibrated measurement of the atomic weight of carbon
\emph{Chin. Sci. Bull.}, \strong{1990}, 35 (290)

Coplen, T., C. Kendall, Hopple J. Comparison of stable isotope reference
samples \emph{Nature}, \strong{1983}, 302 (5905) , pp. 236-238, 10.1038/302236a0

Brand, W. A., Assonov, S. S., & Coplen, T. B. Correction for the
\eqn{^{17}}{17}O interference in \eqn{\delta^{13}}{δ13}C measurements when
analyzing CO\eqn{_{2}}{2} with stable isotope mass spectrometry (IUPAC
Technical Report). \emph{Pure and Applied Chemistry}, \strong{2010}, \emph{82(8)},
1719–1733. \url{https://doi.org/10.1351/PAC-REP-09-01-05}
}
\keyword{internal}
