% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_values.R
\name{delta_values}
\alias{delta_values}
\title{Calculate abundance ratios, δ18O, δ13C, δ's, and raw Δ values}
\usage{
delta_values(dat, d13C_PDB_wg = NULL, d18O_PDBCO2_wg = NULL,
  method = "linterp", genplot = default(genplot),
  quiet = default(quiet), plot_info = NULL,
  plot_column = quo(D47raw))
}
\arguments{
\item{dat}{A \link[tibble:tibble-package]{tibble}, resulting from \code{\link[=correct_backgrounds]{correct_backgrounds()}}.}

\item{method}{Method for matching reference gas to sample gas. Defaults to "linterp".}

\item{plot_info}{The needed metadata for plotting. Result of \code{\link[=clean_did_info]{clean_did_info()}}.}

\item{plot_column}{The quoted column that will be plotted. Defaults to \code{quo(D47raw)}.}

\item{d13C_PBD_wg}{Working gas δ13C value.}

\item{d18O_PBDCO2_wg}{Working gas δ18O value.}
}
\description{
This is a wrapper function that calculates all delta values.
}
