% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_values.R
\name{delta_values}
\alias{delta_values}
\title{Calculate abundance ratios, \eqn{\delta^{18}}O, \eqn{\delta^{13}}C, \eqn{\delta}'s, and raw \eqn{\Delta} values}
\usage{
delta_values(.data, ..., quiet = NULL)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble}, resulting from
\code{\link[=correct_backgrounds]{correct_backgrounds()}}.}
}
\description{
Calculate abundance ratios, \eqn{\delta^{18}}O, \eqn{\delta^{13}}C, \eqn{\delta}'s, and raw \eqn{\Delta} values
}
\details{
This function is a wrapper for the \code{\link[=abundance_ratios]{abundance_ratios()}},
\code{\link[=little_deltas]{little_deltas()}}, and \code{\link[=bulk_and_clumping_deltas]{bulk_and_clumping_deltas()}} functions.
}
