% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_transfer_function.R
\name{empirical_transfer_function}
\alias{empirical_transfer_function}
\title{Empirical Transfer Function}
\usage{
empirical_transfer_function(
  .data,
  ...,
  std_names = paste0("ETH-", 1:3),
  std_values = c(0.258, 0.256, 0.691) - 0.062,
  raw = D47_raw,
  exp = expected_D47,
  session = Preparation,
  id1 = `Identifier 1`,
  etf = etf,
  etf_coefs = etf_coefs,
  slope = slope,
  intercept = intercept,
  out = D47_etf,
  outlier = outlier,
  quiet = NULL,
  parallel = FALSE
)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble} resulting from \code{\link[=collapse_cycles]{collapse_cycles()}}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{std_names}{Names of the standards.}

\item{std_values}{Expected values of the standards. Defaults to Bernasconi et al., 2021.}

\item{raw}{Column name of raw \eqn{\Delta_{47}} values.}

\item{exp}{Name of the new column that will hold expected values.}

\item{session}{The column name to group analyses by. Defaults to
\code{Preparation}.}

\item{id1}{Column name of \verb{Identifier 1} (default).}

\item{etf}{The column name of the new model.}

\item{etf_coefs}{The column name with the coefficients of the model.}

\item{slope}{The column name of the new slope.}

\item{intercept}{The column name of the new intercept.}

\item{out}{The new column name.}

\item{outlier}{Column name of \code{outlier} (default).}

\item{quiet}{If \code{TRUE}, print info messages. If \code{NULL} (default) listens to \code{options(clumpedr.quiet)}}

\item{parallel}{Whether or not (default) to process this in parallel, using package \code{furrr}.}
}
\description{
Calculate and apply the empirical transfer function (ETF). First append the
expected values for the standards, then calculate the ETF per Preparation,
then apply the ETF to all the samples and standards.
}
\details{
This function is a wrapper for \code{\link[=append_expected_values]{append_expected_values()}},
\code{\link[=calculate_etf]{calculate_etf()}}, and \code{\link[=apply_etf]{apply_etf()}}.
}
\seealso{
Other empirical transfer functions: 
\code{\link{append_expected_values}()},
\code{\link{apply_etf}()},
\code{\link{calculate_etf}()}
}
\concept{empirical transfer functions}
