% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_outliers}
\alias{find_outliers}
\title{find_outliers}
\usage{
find_outliers(dat, init = 8000, diff = 1200, nsd_off = 4,
  std_names = paste0("ETH-", 1:3))
}
\arguments{
\item{dat}{A \link[tibble:tibble-package]{tibble}, resulting from \code{\link[=previous_function]{previous_function()}}}

\item{parameter}{parameter description}
}
\description{
This function finds outliers.
}
