% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_outliers}
\alias{find_outliers}
\title{Find outliers}
\usage{
find_outliers(.data, init_low = 8000, init_high = 40000, diff = 1200,
  nsd_off = 4, n_id1 = 5, D47 = D47_raw, std_names = paste0("ETH-",
  1:3), session = Preparation, id1 = `Identifier 1`)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble} with raw Delta values and file information.}

\item{init_low}{Minimum initial intensity threshold for mass 44.}

\item{init_high}{Maximum initial intensity threshold for mass 44.}

\item{diff}{Maximum initial difference in mass 44 threshold between standard and sample gas.}

\item{nsd_off}{Number of standard deviations away from the mean threshold.}

\item{n_id1}{Minimum number of aliquots within session to calculate threshold within group.}

\item{D47}{The column with \eqn{\Delta_{47}}{Δ47} values.}

\item{std_names}{Names of the standards used for the correction.}

\item{session}{Column name that defines correction session.}

\item{id1}{Column name of the sample/standard identifier.}
}
\description{
This function finds outliers based on several criteria.
}
\details{
Here, we define an outlier as a measurement that has:
\itemize{
\item an initial mass 44 intensity below \code{init_low}.
\item an initial mass 44 intensity above \code{init_high}.
\item an imbalance between sample and reference gas mass 44 intensities of more than \code{diff}.
\item a clumped value that is more than \code{nsd_off} standard deviations away from the mean.
}
}
