% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_did_info.R
\name{get_inits}
\alias{get_inits}
\title{Get initial intensities of specified mass}
\usage{
get_inits(.did)
}
\arguments{
\item{.did}{The raw data, resulting from \code{\link[isoreader:iso_get_raw_data]{isoreader::iso_get_raw_data()}}.}
}
\value{
A new \link[tibble:tibble-package]{tibble} with columns file_id, s44_init, and r44_init
}
\description{
Get initial intensities of specified mass
}
\examples{
isoreader::iso_get_raw_data(standards, include_file_info = "Analysis") |>
  get_inits()
}
\seealso{
Other metadata cleaning functions: 
\code{\link{add_info}()},
\code{\link{clean_did_info}()},
\code{\link{parse_info}()}
}
\concept{metadata cleaning functions}
