% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_deltas.R
\name{little_deltas}
\alias{little_deltas}
\title{Calculate little deltas from abundance ratios}
\usage{
little_deltas(.data, ..., quiet = NULL)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble} with abundance ratios.}

\item{...}{These dots are for future extensions and must be empty.}

\item{quiet}{If \code{TRUE}, print info messages. If \code{NULL} (default) listens to \code{options(clumpedr.quiet)}}
}
\value{
Same as \code{.data} but with new columns d45-d49.
}
\description{
For all intensities \eqn{i},
\deqn{\delta_i = (R_i / R_iwg - 1) \times 1000}
}
\examples{
# generate some artificial data to show how it works
x <- function() rnorm(n = 10)
dat <- tibble::tibble(R45 = x(), R46 = x(), R47 = x(), R48 = x(), R49 = x(),
                      R45_wg = x(), R46_wg = x(), R47_wg = x(), R48_wg = x(), R49_wg = x())
dat <- dat |>
   little_deltas()
}
