% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_deltas.R
\encoding{UTF-8}
\name{little_deltas}
\alias{little_deltas}
\title{Calculate little deltas from abundance ratios}
\usage{
little_deltas(dat, quiet = default(quiet))
}
\arguments{
\item{dat}{A dataframe with abundance ratios.}
}
\description{
Based on the $(R_i / R_i_wg - 1) * 1000$ formula.
}
