% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_deltas.R
\name{little_deltas}
\alias{little_deltas}
\title{Calculate little deltas from abundance ratios}
\usage{
little_deltas(.data, ..., quiet = NULL)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble} with abundance ratios.}
}
\description{
For all intensities \eqn{i},
\deqn{\delta_i = (R_i / R_iwg - 1) \times 1000}
}
