% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_match.R
\name{match_intensities}
\alias{match_intensities}
\title{Match the reference gas intensity to that of the sample gas}
\usage{
match_intensities(
  data,
  method = "normal",
  masses = c(44:49, 54),
  quiet = default(quiet)
)
}
\arguments{
\item{data}{A \link[tibble:tibble-package]{tibble} with s44--s49 and r44--r49; output of
\code{\link[=spread_intensities]{spread_intensities()}}.}

\item{method}{"linterp" for linear interpolation, or "normal" for
conventional bracketing of sample gas.}

\item{masses}{The masses to generate r and s columns from.}
}
\description{
This matches the reference gas intensity of mass 44 to the sample gas
intensity of mass 44 through linear interpolation (option \code{method = "linterp"}), and then applies this same offset to the other masses.
}
