% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_did_info.R
\name{parse_info}
\alias{parse_info}
\title{Parse info into appropriate types.}
\usage{
parse_info(.data, masspec, std_names = paste0("ETH-", 1:4),
  oth_name = "other", ms_name = masspec, broadid_name = broadid,
  id1 = `Identifier 1`)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble} resulting from \code{\link[isoreader:iso_get_file_info]{isoreader::iso_get_file_info()}}.}

\item{masspec}{The name of the mass spectrometer to append.}

\item{std_names}{Character vector of the standard names to find in
'Identifier 1'.}

\item{oth_name}{Single general name to assign to non-standard "other" measurements.}

\item{ms_name}{The name of the new masspec column.}

\item{broadid_name}{The name of the new broadid column.}

\item{id1}{The column with sample/standard names.}
}
\description{
Converts the information from \code{\link[isoreader:iso_get_file_info]{isoreader::iso_get_file_info()}} into the
appropriate types.
}
\seealso{
Other metadata cleaning functions: \code{\link{add_info}},
  \code{\link{clean_did_info}}, \code{\link{get_inits}}
}
\concept{metadata cleaning functions}
