% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_did_info.R
\name{parse_info}
\alias{parse_info}
\title{Parse info}
\usage{
parse_info(dat, masspec, std_names = paste0("ETH-", 1:4),
  oth_name = "other")
}
\arguments{
\item{dat}{The result of \code{\link[isoreader:iso_get_file_info]{isoreader::iso_get_file_info()}}.}

\item{masspec}{The name of the mass spectrometer to append.}

\item{std_names}{Character vector of the standard names to find in
'Identifier 1'.}

\item{oth_name}{Single general name to assign to non-standard "other" measurements.}
}
\description{
Converts the information from \code{\link[isoreaser:iso_get_file_info]{isoreaser::iso_get_file_info()}} into the
appropriate types.
}
