% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_base.R
\name{plot_base}
\alias{plot_base}
\title{Create a basic plot layer for further expansion.}
\usage{
plot_base(.data, ...)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble}, resulting from any of the
processing functions that returns the iso_file_info.}

\item{...}{Additional aesthetics to pass to ggplot.}
}
\description{
This function creates a ggplot with all of the metadata applied to named
default aesthetics. This is only useful when creating interactive plot with
\link[plotly:ggplotly]{ggplotly}, since all the metadata aesthetics show up when
hovering the mouse over a point.
}
