% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_delta.R
\name{plot_delta}
\alias{plot_delta}
\title{Create a plot of \eqn{\Delta_{47}}{Δ47} values as a function of measurement
time.}
\usage{
plot_delta(dat, x = file_datetime, y = D47_final, ...)
}
\arguments{
\item{...}{Additional aesthetics to pass to \code{\link[=plot_base]{plot_base()}}.}

\item{.data}{A \link[tibble:tibble-package]{tibble}, resulting from \code{\link[=temperature_calculation]{temperature_calculation()}}.}
}
\description{
Create a plot of \eqn{\Delta_{47}}{Δ47} values as a function of measurement
time.
}
