% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_disabled_cycles.R
\name{plot_disabled_cycles}
\alias{plot_disabled_cycles}
\title{Plot disabled cycles}
\usage{
plot_disabled_cycles(.data, y = v44.mV, min = 1500, max = 50000,
  quiet = default(quiet))
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble} with output from \code{\link[=disable_cycles]{disable_cycles()}}.}

\item{min}{Minimum intensity cutoff.}

\item{max}{Maximum intensity cutoff.}
}
\description{
Plot disabled cycles
}
\seealso{
Other cycle functions: \code{\link{disable_cycles}},
  \code{\link{find_bad_cycles}}
}
\concept{cycle functions}
