% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_etf.R
\name{plot_etf}
\alias{plot_etf}
\title{Plot the Empirical Transfer Function}
\usage{
plot_etf(.data, std_names = paste0("ETH-", 1:3), D47_etf = D47_etf,
  raw = D47_raw, exp = expected_D47, session = Preparation)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble}.}

\item{std_names}{The names of the standards.}

\item{D47_etf}{The column name that holds the \eqn{\Delta_{47}}{Δ47} values
after the application of the empirical transfer function.}

\item{raw}{The column name with raw \eqn{\Delta_{47}}{Δ47} values.}

\item{exp}{The column name with expected \eqn{\Delta_{47}}{Δ47} values.}

\item{session}{The session to facet by. Defaults to \code{Preparation}.}
}
\description{
Create a plot with the expected \eqn{\Delta_{47}}{Δ47} value on the x-axis
and the \eqn{\Delta_{47}}{Δ47} derrived from the ETF on the y-axis.
}
\seealso{
Other empirical transfer functions: \code{\link{append_expected_values}},
  \code{\link{apply_etf}},
  \code{\link{empirical_transfer_function}}
}
\concept{empirical transfer functions}
