% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_delta.R
\name{plot_raw_delta}
\alias{plot_raw_delta}
\title{Plot raw delta values as a function of measurement time.}
\usage{
plot_raw_delta(dat, info, column = quo(D47_raw), rawpoints = FALSE,
  pointalpha = 0.5, quiet = default(quiet))
}
\arguments{
\item{dat}{The input tibble, result of \code{bulk_and_clumping_deltas()}.}

\item{info}{The metadata of the raw cycles, used to get the \code{file_datetime}.}

\item{rawpoints}{Logical, whether to include individual cycle estimates (defaults to \code{FALSE}).}

\item{pointalpha}{The alpha value of the raw points if they are plotted.}
}
\description{
Cycles are averaged and pointranges indicate one standard deviation offset
based on the number of cycles taken into account.
}
\details{
Note that with \eqn{\delta^{13}}C and \eqn{\delta^{18}}O, the
pointrange is invisible because the cycles have so little variance that
the line segment usually falls within the dots.
}
