% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stds.R
\name{plot_stds}
\alias{plot_stds}
\title{Plot standards}
\usage{
plot_stds(.data, x = d47_mean, y = D47_final, ...)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble}, resulting from \code{\link[=acid_fractionation]{acid_fractionation()}}}

\item{x}{Column to put on the x axis.}

\item{y}{Column to put on the y ayis.}

\item{...}{Additional aesthetics to pass to \code{\link[=plot_base]{plot_base()}}.}
}
\description{
Plot the standards in the familiar \eqn{\delta}{δ} vs \eqn{\Delta}{Δ} plot.
}
