% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers}
\usage{
remove_outliers(.data, init_low = 8000, init_high = 40000,
  diff = 1200, nsd_off = 4, std_names = paste0("ETH-", 1:3),
  D47 = D47_raw, plot_x = file_datetime, session = Preparation,
  quiet = default(quiet), genplot = default(genplot))
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble}, resulting from
\code{\link[=collapse_cycles]{collapse_cycles()}}.}

\item{init_low}{Minimum initial intensity threshold for mass 44.}

\item{init_high}{Maximum initial intensity threshold for mass 44.}

\item{diff}{Maximum initial difference in mass 44 threshold between standard and sample gas.}

\item{nsd_off}{The number of standard deviations away from the median
Preparation of the standards.}

\item{std_names}{Names of the standards used for the correction.}

\item{D47}{The column with \eqn{\Delta_47}{Δ47} values. Defaults to \code{D47_raw_mean}.}

\item{plot_x}{The column to use for plotting the x axis.}

\item{session}{Column name that defines correction session.}
}
\description{
Remove outliers
}
\details{
This function removes outliers based on several criteria. 1) The
initial intensity of both sample and reference gas is above 8 V. 2) The
difference in intensity between sample and reference gas is less than 1.2
V. 3) The sample or reference raw \eqn{\Delta_47}{Δ47} value is less than
4 SD away from the preparation mean.
}
