% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers}
\usage{
remove_outliers(.data, init = 8000, diff = 1200, nsd_off = 4,
  std_names = paste0("ETH-", 1:3), plot_x = file_datetime,
  plot_y = D47_raw_mean, session = Preparation,
  quiet = default(quiet), genplot = default(genplot))
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble}, resulting from
\code{\link[=collapse_cycles]{collapse_cycles()}}.}

\item{init}{The minimum initial intensity of mass 44.}

\item{diff}{The maximum initial difference in intensity of mass 44.}

\item{nsd_off}{The number of standard deviations away from the median
Preparation of the standards.}

\item{std_names}{Names of the standards used for the correction.}

\item{plot_x}{The column to use for plotting the x axis.}

\item{plot_y}{The column to use for plotting the y axis. Defaults to
\code{D47_raw_mean}.}

\item{session}{Column name that defines correction session.}
}
\description{
Remove outliers
}
\details{
This function removes outliers based on several criteria. 1) The
initial intensity of both sample and reference gas is above 8 V. 2) The
difference in intensity between sample and reference gas is less than 1.2
V. 3) The sample or reference raw \eqn{\Delta_47} value is less than 4 SD
away from the preparation mean.
}
