% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers}
\usage{
remove_outliers(
  .data,
  init_low = 8000,
  init_high = 40000,
  init_diff = 1200,
  param49_off = 1,
  internal_sd = 0.15,
  n_min = 5,
  n_id1 = 5,
  nsd_off = 4,
  D47 = D47_raw,
  session = Preparation,
  id1 = `Identifier 1`,
  plot_x = file_datetime,
  quiet = default(quiet),
  genplot = default(genplot)
)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble}, resulting from
\code{\link[=collapse_cycles]{collapse_cycles()}}.}

\item{init_low}{Minimum initial intensity threshold for mass 44.}

\item{init_high}{Maximum initial intensity threshold for mass 44.}

\item{init_diff}{Maximum initial difference in mass 44 threshold between standard and sample gas.}

\item{param49_off}{Where param_49 > this value is considered an outlier.}

\item{internal_sd}{The internal standard deviation outlier cutoff.}

\item{n_min}{Minimum number of aliquots within session to calculate threshold.}

\item{n_id1}{Minimum number of aliquots within session to calculate threshold within group.}

\item{nsd_off}{The number of standard deviations away from the median
Preparation of the standards.}

\item{D47}{The column with \eqn{\Delta_47}{Δ47} values. Defaults to \code{D47_raw_mean}.}

\item{session}{Column name that defines correction session.}

\item{id1}{Column name of the sample/standard identifier.}

\item{plot_x}{The column to use for plotting the x axis.}
}
\description{
Remove outliers
}
\details{
This function removes outliers based on several criteria. 1) The
initial intensity of both sample and reference gas is above 8 V. 2) The
difference in intensity between sample and reference gas is less than 1.2
V. 3) The sample or reference raw \eqn{\Delta_47}{Δ47} value is less than
4 SD away from the preparation mean.
}
