% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_match.R
\name{spread_match}
\alias{spread_match}
\title{spread_match}
\usage{
spread_match(dat, method = "normal", quiet = default(quiet))
}
\arguments{
\item{dat}{a data frame or tibble containing mass intensities per cycle}

\item{method}{"linterp" for linear interpolation, or "normal" for
conventional bracketing of sample gas.}
}
\description{
Spread and match sample gas and reference gas.
}
