% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_match.R
\name{spread_match}
\alias{spread_match}
\title{Spread and match sample gas and reference gas.}
\usage{
spread_match(
  .data,
  ...,
  method = "normal",
  masses = c(44:49, 54),
  quiet = NULL
)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble} with s44--s49 and r44--r49; output of
\code{\link[=spread_intensities]{spread_intensities()}}.}

\item{method}{"linterp" for linear interpolation, or "normal" for
conventional bracketing of sample gas.}

\item{masses}{The masses to generate r and s columns from.}
}
\description{
Spread and match sample gas and reference gas.
}
