% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standards}
\alias{standards}
\title{Several standard did files}
\format{
An object of class \code{iso_file_list} of length 27.
}
\usage{
standards
}
\description{
Several standard did files
}
\keyword{datasets}
