% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_calculation.R
\name{temperature_calculation}
\alias{temperature_calculation}
\title{Calculate the temperatures}
\usage{
temperature_calculation(
  .data,
  D47 = D47_final,
  temp = temperature,
  quiet = default(quiet)
)
}
\arguments{
\item{.data}{A \link[tibble:tibble-package]{tibble}, resulting from \code{\link[=acid_fractionation]{acid_fractionation()}}.}

\item{D47}{The column name of the \eqn{\Delta_47} values to use as input.}

\item{temp}{The column name of the output temperature.}
}
\description{
This uses the \code{\link[=revcal]{revcal()}} function to calculate temperatures from
\eqn{\Delta_{47}}{Δ47} values. At the moment it ignores the uncertainty in
the regression.
}
\seealso{
revcal tempcal
}
