% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_calculation.R
\name{temperature_calculation}
\alias{temperature_calculation}
\title{Calculate the temperatures}
\usage{
temperature_calculation(dat, D47 = quo(D47_final),
  quiet = default(quiet))
}
\arguments{
\item{dat}{A \link[tibble:tibble-package]{tibble}, resulting from \code{acid_fractionation()}}

\item{D47}{The quoted column name of the \eqn{\Delta_47} values to use as input.}
}
\description{
Calculate the temperatures
}
\seealso{
revcal tempcal
}
