% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_calculation.R
\name{temperature_calculation}
\alias{temperature_calculation}
\title{temperature_calculation}
\usage{
temperature_calculation(dat, quiet = default(quiet))
}
\arguments{
\item{dat}{A \link[tibble:tibble-package]{tibble}, resulting from \code{acid_fractionation()}}
}
\description{
This calculates the temperatures.
}
